added agent UI
