+++ ui/mod/List.js
@@ -1,8 +1,9 @@
-function populate(self,coll,Row){
+function populate(self, coll, Row, params){
 	const ids=Object.keys(coll.models)
-	for(let i=0,k; (k=ids[i]); i++){
+	for(let i=0,l=ids.length,k; i<l; i++){
+		k = ids[i]
 		// TODO: how to pass model directly wo params
-		self.spawn(Row,{id: k},[['model','model','coll','id']])
+		self.spawn(Row,Object.assign({id: k}, params),[['model','model','coll','id']])
 	}
 }
 
@@ -11,22 +12,21 @@ return {
 		coll:'models',
 		Row:'view'
 	},
-	create:function(deps){
+	create:function(deps, params){
 		deps.coll.callback.on('add',function(evt, coll, id){
 			// eslint-disable-next-line
 			console.log('Coll.add',evt, coll.get(id))
-			this.spawn(deps.Row,{id},[['model','model','coll','id']])
+			this.spawn(deps.Row,Object.assign({id}, params),[['model','model','coll','id']])
 		},this)
 		deps.coll.callback.on('update',function(){
 			// eslint-disable-next-line
 			console.log('Coll.update',arguments)
 		},this)
 
-		populate(this, deps.coll, deps.Row)
+		populate(this, deps.coll, deps.Row, params)
 	},
 	remove:function(){
 		const coll=this.deps.coll
-		coll.models[1].callback.off(null,null,this)
 		coll.callback.off(null,null,this)
 		this.super.remove.call(this)
 	}
