+++ ui/mod/Scrollable.js
@@ -1,17 +1,7 @@
-const router = require('po/router')
+// A scrollable container
 
 return {
-	deps: {
-		ghSession: 'session',
-		aud: 'list'
-	},
 	create(deps, params){
-		const cred = deps.ghSession.get('cred')
-		if (cred && cred.access_token){
-			const bpayload = cred.access_token.split('.')[1]
-			const payload = JSON.parse(atob(bpayload))
-			deps.aud.push(...payload.aud.split(','))
-		}
 		this.super.create.call(this, deps, params)
 	}
 }
